backend/project/requirements.txt
asgiref==3.7.2
attrs==23.1.0
blinker==1.6.2
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.4
colorama==0.4.6
crashtest==0.4.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dulwich==0.21.5
filelock==3.12.2
Flask==2.3.2
html5lib==1.1
httpie==3.2.2
idna==3.4
importlib-metadata==6.7.0
https://cdn.mysql.com/Downloads/Connector-Python/mysql-connector-python-2.1.3.tar.gz
